Use [ instead of [[ for question 20.4.6
At the end of section 20.4, Hadley writes

There is an important variation of [ called [[. [[ only ever extracts a single element, and always drops names. It’s a good idea to use it whenever you want to make it clear that you’re extracting a single item, as in a for loop. The distinction between [ and [[ is most important for lists, as we’ll see shortly.

I think [ should be used here.