Fix answer to Ex. 16.4.5.5
The original answer said it was fine. The answer should be that when dividing an interval by a period use `%/%`. However, the example doesn't raise a warning, and seems to produce the "correct" answer (12 months). I'm not exactly sure why.

Fixes #156

There is also a typo in the original question, so I'm going to open a PR in the r4ds repo.